Move JsMethod call stuff into separate file
Refactor to reduce diff in #4699
